test: passed #get_input method testing
In this test, I used a new matcher method called #have_received.
This was required because I was testing a puts message that would have come 3rd or 4th order during execution.
With this new matcher, I could check whether the method had called puts with the so-and-so error message during the execution? Without the need to code and test every other error message sequentially.